day 8
This is probably trivial if you have better graph theory than I do. I
wasted a bunch of time thinking this was a "cliques" problem before I
realised the reason the numbers looked wrong is because cliques are
something different.. and then I had to trawl through the networkx
documentation looking at the graph-theory definitons of things until I
found the one I wanted. And then it only took me another 7 minutes to
get it working for part 2 which could well be the best comeback ratio
I'll ever see